+++ database/migrations/2019_06_10_062653_create_tasks_table.php
@@ -18,10 +18,10 @@ class CreateTasksTable extends Migration
 
             $table->mediumText('description');
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
 
-            $table->integer('event_id')->unsigned();
+            $table->bigInteger('event_id')->unsigned();
             $table->foreign('event_id')->references('id')->on('events');
 
             $table->integer('is_approved')->default(0);
