+++ database/migrations/2019_06_10_065009_create_socialmedia_accounts_table.php
@@ -18,8 +18,9 @@ class CreateSocialmediaAccountsTable extends Migration
             $table->string('platform');
             $table->string('username');
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
+
             $table->timestamps();
         });
     }
