+++ database/migrations/2019_06_10_072722_create_total_points_table.php
@@ -14,9 +14,9 @@ class CreateTotalPointsTable extends Migration
     public function up()
     {
         Schema::create('total_points', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
 
             $table->integer('value')->default(0);
