+++ database/migrations/2019_06_10_073019_create_unapproved_images_table.php
@@ -14,10 +14,11 @@ class CreateUnapprovedImagesTable extends Migration
     public function up()
     {
         Schema::create('unapproved_images', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-           $table->integer('user_id')->unsigned();
+           $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
+            
             $table->integer('is_approved')->default(0);
             $table->timestamps();
         });
